added comments to broker
The body parser middleware is used to parse the data from the HTTP request body. It's a crucial component because it allows our server to extract and work with data sent by clients in the body of HTTP requests, such as data submitted through HTML forms or data sent as JSON.

Here's why the body parser is important:

Parsing Data Formats: HTTP requests can contain data in various formats, including JSON, form data, XML, and more. The body parser middleware is responsible for detecting the format of the incoming data and parsing it into a JavaScript object that your application can work with.

Accessing POST Data: When clients submit data through HTML forms with the POST method, that data is included in the request body. Without the body parser, you wouldn't be able to access this data in a structured manner.

Handling JSON Data:  The body parser can parse incoming JSON data into JavaScript objects, making it easy to work with JSON payloads in your application.

Middleware Chain: The body parser is often used early in the middleware chain to ensure that data is parsed and available for subsequent route handlers. This allows you to process / validate data before passing it on to the relevant route handler.